requirements.txt
aioredis==1.3.1
amqp==2.6.1
anyascii==0.2.0
asgiref
astroid==2.4.1
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.7.1
Automat==20.2.0
autopep8==1.5.4
bcrypt==3.1.7
beautifulsoup4==4.9.3
billiard==3.6.4.0
blessed==1.17.6
boto3==1.16.10
botocore==1.19.10
cached-property==1.5.1
celery==4.4.2
cement==2.8.2
certifi==2020.4.5.1
cffi==1.14.0
channels==2.4.0
channels-redis==3.1.0
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.3
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.9.2
cssutils==1.0.2
daphne==2.5.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.2.6
django-allauth==0.42.0
django-autoslug==1.9.8
django-appconf==1.0.4
django-cors-headers==3.3.0
django-cryptography==1.0
django-celery-beat
django-debug-toolbar==2.2
django-decouple==2.1
django-filter==2.3.0
django-hosts==4.0
django-inlinecss==0.3.0
django-modelcluster==5.1
django-oauth-plus==2.2.9
django-oauth2-provider==0.2.6.1
django-rest-auth==0.9.5
django-ses==1.0.3
django-split-settings==1.0.1
django-storages==1.10.1
django-summernote==0.8.11.6
django-taggit==1.4.0
django-treebeard==4.5.1
djangorestframework==3.12.4
djangorestframework-oauth==1.1.0
djangorestframework-simplejwt==4.8.0
docker==4.2.0
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
future==0.16.0
gunicorn==20.0.4
hiredis==1.1.0
html5lib==1.1
httplib2==0.18.1
hyperlink==20.0.1
idna==2.7
incremental==17.5.0
isort==4.3.21
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.10.0
jsonfield==3.1.0
jsonschema==3.2.0
kombu==4.6.11
l18n==2020.6.1
lazy-object-proxy==1.4.3
Markdown==3.2.2
MarkupSafe==1.1.1
mccabe==0.6.1
msgpack==1.0.0
numpy==1.19.4
oauth2==1.9.0.post1
oauthlib==3.1.0
openpyxl==3.0.7
paramiko==2.7.1
pathlib==1.0.1
pathspec==0.5.9
Pillow==7.1.2
prompt-toolkit==3.0.18
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==2.1.0
PyNaCl==1.4.0
pynliner==0.8.0
pyOpenSSL==19.1.0
pyrsistent==0.16.0
python-dateutil==2.8.0
python-decouple==3.3
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.3.1
redis==3.5.3
requests==2.24.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
semantic-version==2.5.0
service-identity==18.1.0
shortuuid==1.0.1
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
soupsieve==2.0.1
sqlparse==0.3.1
tablib==3.0.0
telepath==0.2
termcolor==1.1.0
texttable==1.6.2
toml==0.10.1
twilio==6.48.0
Twisted==20.3.0
txaio==20.4.1
uritemplate==3.0.1
urllib3==1.25.11
vine==1.3.0
wagtail==2.13.1
wcwidth==0.1.9
webencodings==0.5.1
websocket-client==0.57.0
Willow==1.4
wrapt==1.12.1
xlrd==2.0.1
XlsxWriter==1.4.3
xlwt==1.3.0
zope.interface==5.1.2
